改用ShortArray来接收音频数据

+++ minerva/src/main/java/com/lodz/android/minerva/bean/states/VadDetect.kt
@@ -5,4 +5,4 @@ package com.lodz.android.minerva.bean.states
  * @author zhouL
  * @date 2022/11/10
  */
-class VadDetect(data: ByteArray?, end: Int, val isSpeech: Boolean) : Recording(data, end)
+class VadDetect(data: ShortArray?, end: Int, val isSpeech: Boolean) : Recording(data, end)
